MB-68266: Fix narrowing warnings in io
Use gsl::narrow<uint32_t> when passing size_t to setBodylen() in
read_callback_test.cc. The MCBP protocol defines bodylen as uint32_t,
but the test helper function takes size_t for convenience.

Using gsl::narrow instead of narrow_cast ensures test values are in
valid range with runtime checks rather than silently truncating.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/244289
Tested-by: Build Bot <build@couchbase.com>

+++ io/network/read_callback_test.cc
@@ -10,6 +10,7 @@
 
 #include "read_callback.h"
 #include <folly/portability/GTest.h>
+#include <gsl/gsl-lite.hpp>
 #include <mcbp/protocol/opcode.h>
 #include <mcbp/protocol/request.h>
 #include <memcached/vbucket.h>
@@ -42,7 +43,7 @@ protected:
         request->setVBucket(Vbid{666});
         request->setOpcode(cb::mcbp::ClientOpcode::Noop);
         request->setCas(0xdeadbeef);
-        request->setBodylen(bodylen);
+        request->setBodylen(gsl::narrow<uint32_t>(bodylen));
         addDataToStream(request->getFrame());
     }
